+++ src/renderer/Layout.tsx
@@ -1,5 +1,11 @@
 import * as React from 'react';
-import { Pages, RecStatus, AppState, RendererVideo } from 'main/types';
+import {
+  AdvancedLoggingStatus,
+  Pages,
+  RecStatus,
+  AppState,
+  RendererVideo,
+} from 'main/types';
 import { Dispatch, MutableRefObject, SetStateAction } from 'react';
 import { ConfigurationSchema } from 'config/configSchema';
 import SceneEditor from './SceneEditor';
@@ -16,6 +22,9 @@ interface IProps {
   playerHeight: MutableRefObject<number>;
   config: ConfigurationSchema;
   setConfig: Dispatch<SetStateAction<ConfigurationSchema>>;
+  advancedLoggingStatus: AdvancedLoggingStatus;
+  previewEnabled: boolean;
+  setPreviewEnabled: Dispatch<SetStateAction<boolean>>;
 }
 
 /**
@@ -32,6 +41,9 @@ const Layout = (props: IProps) => {
     playerHeight,
     config,
     setConfig,
+    advancedLoggingStatus,
+    previewEnabled,
+    setPreviewEnabled,
   } = props;
   const { page, category } = appState;
 
@@ -57,6 +69,7 @@ const Layout = (props: IProps) => {
         setConfig={setConfig}
         appState={appState}
         setAppState={setAppState}
+        advancedLoggingStatus={advancedLoggingStatus}
       />
     );
   };
@@ -68,6 +81,8 @@ const Layout = (props: IProps) => {
         appState={appState}
         config={config}
         setConfig={setConfig}
+        previewEnabled={previewEnabled}
+        setPreviewEnabled={setPreviewEnabled}
       />
     );
   };