+++ src/renderer/containers/ApplicationStatusCard/Status.tsx
@@ -1,9 +1,15 @@
 import { getLocalePhrase } from 'localisation/translations';
 import { Phrase } from 'localisation/phrases';
-import { HardDriveDownload } from 'lucide-react';
+import { HardDriveDownload, TriangleAlert } from 'lucide-react';
 import { ConfigurationSchema } from 'config/configSchema';
-import { ActivityStatus, AppState, RecStatus, SaveStatus } from 'main/types';
-import React, { useEffect, useState } from 'react';
+import {
+  ActivityStatus,
+  AdvancedLoggingStatus,
+  AppState,
+  RecStatus,
+  SaveStatus,
+} from 'main/types';
+import React, { Dispatch, SetStateAction, useEffect, useState } from 'react';
 import { Button } from 'renderer/components/Button/Button';
 import {
   HoverCard,
@@ -11,6 +17,7 @@ import {
   HoverCardTrigger,
 } from 'renderer/components/HoverCard/HoverCard';
 import Separator from 'renderer/components/Separator/Separator';
+import { Tooltip } from 'renderer/components/Tooltip/Tooltip';
 import StatusLight, {
   StatusLightProps,
 } from 'renderer/components/StatusLight/StatusLight';
@@ -30,6 +37,8 @@ type StatusProps = {
   config: ConfigurationSchema;
   appState: AppState;
   activityStatus: ActivityStatus | null;
+  advancedLoggingStatus: AdvancedLoggingStatus;
+  setPreviewEnabled: Dispatch<SetStateAction<boolean>>;
 };
 
 const Status = ({
@@ -39,6 +48,8 @@ const Status = ({
   config,
   appState,
   activityStatus,
+  advancedLoggingStatus,
+  setPreviewEnabled,
 }: StatusProps) => {
   const { language } = appState;
   const [recTimerSec, setRecTimerSec] = useState(0);
@@ -198,13 +209,27 @@ const Status = ({
           {getLocalePhrase(language, Phrase.StatusDescriptionWatchingLogs)}
           {': '}
         </p>
-        <ul className="text-xs text-popover-foreground/60 list-disc pl-4">
+        <ul className="text-xs text-popover-foreground/60 list-disc pl-4 whitespace-nowrap">
           {config.recordRetail && (
             <li>
               <span className="font-bold">
                 {getLocalePhrase(language, Phrase.Retail)}
                 {': '}
               </span>
+              {!advancedLoggingStatus.retail && (
+                <Tooltip
+                  content={getLocalePhrase(
+                    language,
+                    Phrase.AdvancedCombatLoggingDisabledWarning,
+                  )}
+                >
+                  <TriangleAlert
+                    size={12}
+                    fill="#facc15"
+                    className="text-yellow-900 mr-1 inline align-middle -mt-0.5"
+                  />
+                </Tooltip>
+              )}
               <code>{config.retailLogPath}</code>
             </li>
           )}
@@ -214,6 +239,20 @@ const Status = ({
                 {getLocalePhrase(language, Phrase.Classic)}
                 {': '}
               </span>
+              {!advancedLoggingStatus.classic && (
+                <Tooltip
+                  content={getLocalePhrase(
+                    language,
+                    Phrase.AdvancedCombatLoggingDisabledWarning,
+                  )}
+                >
+                  <TriangleAlert
+                    size={12}
+                    fill="#facc15"
+                    className="text-yellow-900 mr-1 inline align-middle -mt-0.5"
+                  />
+                </Tooltip>
+              )}
               <code>{config.classicLogPath}</code>
             </li>
           )}
@@ -223,6 +262,20 @@ const Status = ({
                 {getLocalePhrase(language, Phrase.Era)}
                 {': '}
               </span>
+              {!advancedLoggingStatus.era && (
+                <Tooltip
+                  content={getLocalePhrase(
+                    language,
+                    Phrase.AdvancedCombatLoggingDisabledWarning,
+                  )}
+                >
+                  <TriangleAlert
+                    size={12}
+                    fill="#facc15"
+                    className="text-yellow-900 mr-1 inline align-middle -mt-0.5"
+                  />
+                </Tooltip>
+              )}
               <code>{config.eraLogPath}</code>
             </li>
           )}
@@ -232,6 +285,20 @@ const Status = ({
                 {getLocalePhrase(language, Phrase.RetailPtr)}
                 {': '}
               </span>
+              {!advancedLoggingStatus.retailPtr && (
+                <Tooltip
+                  content={getLocalePhrase(
+                    language,
+                    Phrase.AdvancedCombatLoggingDisabledWarning,
+                  )}
+                >
+                  <TriangleAlert
+                    size={12}
+                    fill="#facc15"
+                    className="text-yellow-900 mr-1 inline align-middle -mt-0.5"
+                  />
+                </Tooltip>
+              )}
               <code>{config.retailPtrLogPath}</code>
             </li>
           )}
@@ -241,6 +308,20 @@ const Status = ({
                 {getLocalePhrase(language, Phrase.ClassicPtr)}
                 {': '}
               </span>
+              {!advancedLoggingStatus.classicPtr && (
+                <Tooltip
+                  content={getLocalePhrase(
+                    language,
+                    Phrase.AdvancedCombatLoggingDisabledWarning,
+                  )}
+                >
+                  <TriangleAlert
+                    size={12}
+                    fill="#facc15"
+                    className="text-yellow-900 mr-1 inline align-middle -mt-0.5"
+                  />
+                </Tooltip>
+              )}
               <code>{config.classicPtrLogPath}</code>
             </li>
           )}
@@ -294,7 +375,12 @@ const Status = ({
   const isSaving = savingStatus === SaveStatus.Saving;
 
   return (
-    <HoverCard openDelay={300}>
+    <HoverCard
+      openDelay={300}
+      onOpenChange={(open) => {
+        if (statusDescription) setPreviewEnabled(!open);
+      }}
+    >
       <HoverCardTrigger>
         <div className="w-full h-full flex relative rounded-md border-t border-[rgba(255,255,255,10%)] hover:cursor-pointer">
           <StatusLight
@@ -326,7 +412,7 @@ const Status = ({
           </div>
         </div>
         {!!statusDescription && (
-          <HoverCardContent className="w-[260px] mx-4">
+          <HoverCardContent className="w-fit mx-4">
             {statusDescription}
           </HoverCardContent>
         )}